Got initial PINN code running.
I decided to split the code into two files: one for the Neural Network (
the class itself, the loss functions, the driving term, the residual
function) and another for the learning algorithm (training + testing
loop, data generation, the main function).

I was able to get the code to run (and it appears to be working!).
Currently, I am not batching the boundary/colocation points. I want to
look into doing that in the future.

I also do not have any plotting stuff. I may write a third file that
contains plotting functions and such.

Anywho, I have not verified that my code actually works. I simply got
it to run without errors + it appears to be doing the right thing.
Expect extensive tweaks in the future, as well as possibly a third file
specifically for plotting stuff.