Added network saving + plot colorbars
I added a line of code to save the network's state dictionary to file
after training. My hope is that, in the future, I will have options
to load a network from file (rather than reinvent the network from
scratch). ideally, all of this will be controlled by a setup file.

I also added colorbars to my plots, which stupendously difficult. *Start
of Rant* Adding colorbars to subplots in matplotlib is necessarily
difficult. You have to manually set a half dozen parameters, or the
colorbars will look like absolute garbage. I can not believe how
difficult it was to add the colorbars in a way that did not ruin the
plot. I hope and pray that the developers improve this aspect of the
library, because it's probably one of the most needlessly cumbersome
processes I've ever had to deal with in a major library. Get your crap
together. *end of rant*