Added setup file reader
I wrote a setup file reader, and modified main to work off of the
data read in by this reader rather than hard-set parameters. In theory,
this should allow me to do almost all of my configuration/setting things
from the setup file without having to touch the code.

I also made minor tweaks throughout the code.

I may want to expand the settings in the setup file. For example, it
may be useful to set the optimizer and plot settings from the setup
file. If I find that I tweak these or other settings a lot, I'll add
them to the setup file.

I just realized that I forgot to add the setup file or setup file reader
to this commit. Oops. To be continuted...