Implemented loading + added optimizer saving
I made some tweaks to the save/load stuff. I changed the save routine
so that it saves the state of the optimizer as well as the network. I
changed the loading interface so that there are distinct settings for
loading the network and optimizer (in case you want to load one but
not the other).

If you load the optimizer state, then the learning rate in the setup
file is ignored (instead, we use the saved learning rate). I may change
this in the future, idk.

Anywho, the loading/saving stuff appears to work now and is sufficiently
flexible.