Added code to teach SciML in Torch to Jacob
I split the code in a pretty standard way: There's a network file which houses the neural network class... there's a Losses file which houses the loss functions (as well as "PDE_Residual". I use this for plotting). There's also a Test_Train file which contains the test/train functions. Then there's main, whcih drives everything (sets up a network, optimizer, and data, then runs the epochs, then prints).

I also added a separate file for plotting. This is literally just a boot-strapped version of Plotter in the Robert sub-directory. I cleaned it up a bit, but mostly just copied the old code because I wanted this thing to be done. It seems to work (it plots the soluition + PDE residual on a uniform grid over the problem domain).

The code appears to work / is ready for presentation. I wrote all of this code from scratch (except for the plotter.) I could have used the code I have in the Robert sub-directory, but my understanding of torch has changed so mcuh since I wrote that code that I felt it was better to start over.